requirements.txt
celery==5.2.3
gunicorn==20.1.0
psycopg2==2.9.3
bcrypt==3.2.0
cryptography==3.4.7
Django==3.2.13
django-crispy-forms==1.12.0
volatility3==2.0.1
yara-python==4.1.3
pycryptodome
pefile==2021.9.3
capstone==4.0.2
redis==4.1.0
django-cookiebanner==0.2.5
pyyaml==6.0